+++ src/components/RCollapseGrid/index.ts
@@ -1,3 +1,5 @@
 import RCollapseGrid from './src/index'
+import props from './src/props'
 
 export default RCollapseGrid
+export { props }