+++ src/components/RMoreDropdown/index.ts
@@ -1,0 +1,5 @@
+import RMoreDropdown from './src/index'
+import props from './src/props'
+
+export default RMoreDropdown
+export { props }